requirements.txt
streamlit==1.28.0
pymilvus==2.3.5
pillow==10.0.0
python-dotenv==1.0.0
huggingface_hub>=0.19.0
requests>=2.31.0
pymupdf>=1.23.0
pytesseract>=0.3.10
camelot-py>=0.11.0
pypdf>=3.17.0